Updates to changes in goChem API, bugfixes and made goReduce plugin work
Minor bug fixes in the previous update to new goChem API, modified the
calles to comply with the new goChem API for reading/writing PDBs.

Also used a workaround to make goReduce work, now the pluging transforms
the protonated chempy object to a PDB string, and then reads that PDB
into PyMOL so it is correctly displayed.